did the same with ammo

+++ FoxholeArmorPenetrationCalculator/Tank.cs
@@ -4,7 +4,7 @@
     {
         public string name;
         public double basePenetrationChance;
-        public void DeclareTank(string name, int basePenetrationChance)
+        public void DeclareTank(string name, double basePenetrationChance)
         {
             this.name = name;
             this.basePenetrationChance = basePenetrationChance;
